+++ proton/vpn/app/gtk/assets/release_notes.md
@@ -1,2 +1,6 @@
+## 4.0.0-a15
+- Pinned Tray Connections: with this new setting you can now pin you frequently used connections and access them from system tray for quick connectivity.
+- Moderate NAT: with this setting you can disable randomization of the local addresses mapping for greater gaming performances.
+
 ## 4.0.0-a13
 - We added a new settings window to the app. You can now enable port forwarding, auto-connect, and other VPN features in just a few clicks.